requirements.txt
pystray
Pillow
keyboard
hidapi